Move HotReloadMethodOutcome and its kind enum out of the result file
HotReloadOrchestratorResult.cs held three unrelated types, so a reader
looking for the per-method outcome had to open the run-level aggregate.
Each type now lives in a file named after it; the bodies, namespace and
accessibility are unchanged, so this is a pure move.
