+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadMethodOutcome.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2c192402ab9554dacb11d113cd5522c0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
