+++ Packages/src/Editor/FirstPartyTools/HotReload/HotReloadMethodOutcomeKind.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 851a59e35c8a74b5fb16a7e3f6ab7395
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
